server/controllers/authController.ts
import { Request, Response } from "express";
import { registerUser, loginUser } from "../services/authService";

export const register = async (req: Request, res: Response) => {
  try {
    const user = await registerUser(req.body);
    res.status(201).json({ message: "User registered successfully", user });
  } catch (error) {
    if (error instanceof Error) {
      res.status(400).json({ message: error.message });
    } else {
      res.status(400).json({ message: "An unknown error occurred" });
    }
  }
};

export const login = async (req: Request, res: Response): Promise<void> => {
  try {
    const { email, password } = req.body;
    if (!email || !password) {
      res.status(400).json({ message: "Missing email or password" });
      return;
    }

    // Login the user and get token
    const { user, token } = await loginUser(email, password);

    // Send both token and user information to the frontend
    res.status(200).json({ message: "Login successful", user, token });
  } catch (error) {
    res.status(401).json({ message: error instanceof Error ? error.message : "Authentication failed" });
  }
};